refactor(hero): address Task 2 code-review issues
Five issues from code review on commit 8fc3c62:

Important:
1. Extract runBoot to lib/boot-animation.ts. HeroBootAnimation.tsx
   was 358 lines (43% over the 250 cap); the pure animation driver
   has no React dependency and belongs in lib/. Island now ~100
   lines, just the React wiring (useEffect + matchMedia gate + JSX).
   sysfail-loop test updated to import from the new location.
2. Fix latent ctrl temporal dead zone in onFirstLoop. Previously
   captured ctrl before runBoot assigned it; safe today because the
   callback fires seconds later, but a refactor could silently break
   it. Now uses a stable ctrlRef holder so the closure accesses
   ctrlRef.current at invocation time instead of a pre-assignment
   reference.
3. Collapse the .hero--desktop CSS double-rule. Removed the
   display: block block (only there to satisfy the source-grep test);
   the display: flex rule is now the single source of truth.
   hero-rsc.test.ts regex relaxed to (block|flex) to accept the
   layout-correct rule.

Minor:
4. Extract sysfail timeout magic numbers to named constants
   (SYSFAIL_SHAKE_FRAME_1_MS etc.) near existing timing objects.
5. Remove the thin buildBlankLine wrapper; call sites use
   buildLine([' ']) directly.

No behavior change. All 63 tests still pass.
