+++ components/client/HeroBootAnimation.tsx
@@ -1,238 +1,23 @@
 'use client';
 
 import { useEffect, useRef } from 'react';
+import {
+  type BootCtrl,
+  buildDesktopOnFirstLoop,
+  buildLine,
+  buildStaticCmdLine,
+  buildStaticDialogLine,
+  DESKTOP_DIALOG,
+  DESKTOP_LINE_SPECS,
+  MOBILE_DIALOG,
+  MOBILE_LINE_SPECS,
+  runBoot,
+} from '@/lib/boot-animation';
 import { readMotion } from '@/lib/motion';
 
-// ── Typed line-spec data ──────────────────────────────────────────────────────
-type Span = { cls: string; text: string };
-type LinePart = string | Span;
-
-const DESKTOP_LINE_SPECS: LinePart[][] = [
-  ['[SYSTEM BOOT SEQUENCE INITIATED]'],
-  [' '],
-  ['Initializing kernel modules... ', { cls: 'boot__ok', text: 'OK' }],
-  ['Mounting local filesystems... ', { cls: 'boot__ok', text: 'OK' }],
-  ['Starting network services... ', { cls: 'boot__ok', text: 'OK' }],
-  ['Loading security protocols... ', { cls: 'boot__enc', text: '[ENCRYPTED]' }],
-  [{ cls: 'boot__welcome', text: 'Welcome to DEV_OS v2.0.4-stable [user: erik]' }],
-  [' '],
-];
-
-const MOBILE_LINE_SPECS: LinePart[][] = [
-  ['[BOOT SEQUENCE INITIATED]'],
-  [' '],
-  ['kernel modules... ', { cls: 'boot__ok', text: 'OK' }],
-  ['mount fs... ', { cls: 'boot__ok', text: 'OK' }],
-  ['network... ', { cls: 'boot__ok', text: 'OK' }],
-  ['security... ', { cls: 'boot__enc', text: '[ENCRYPTED]' }],
-  [{ cls: 'boot__welcome', text: 'DEV_OS v2.0.4 [user: erik]' }],
-  [' '],
-];
-
-const DESKTOP_DIALOG = [
-  'Wake up, Neo...',
-  'Wake up...',
-  'The Matrix has you...',
-  'Knock, knock, Neo...',
-];
-const MOBILE_DIALOG = ['Wake up, Neo...', 'The Matrix has you...', 'Knock, knock, Neo...'];
-
-// ── Safe DOM builders (no innerHTML) ─────────────────────────────────────────
-function buildLine(parts: LinePart[]): HTMLElement {
-  const line = document.createElement('span');
-  line.className = 'boot__line';
-  for (const p of parts) {
-    if (typeof p === 'string') {
-      line.appendChild(document.createTextNode(p));
-    } else {
-      const s = document.createElement('span');
-      s.className = p.cls;
-      s.textContent = p.text;
-      line.appendChild(s);
-    }
-  }
-  return line;
-}
-
-function buildBlankLine(): HTMLElement {
-  return buildLine([' ']);
-}
-
-function buildStaticCmdLine(): HTMLElement {
-  return buildLine([
-    { cls: 'boot__prompt', text: 'erik@portfolio:~$' },
-    ' ',
-    { cls: 'boot__cmd', text: 'run bio.exe --verbose' },
-  ]);
-}
-
-function buildStaticDialogLine(text: string): HTMLElement {
-  return buildLine([
-    { cls: 'boot__matrix-prefix', text: '>' },
-    { cls: 'boot__matrix-out', text: text },
-  ]);
-}
-
-// ── Boot animation (DOM mutations, no per-char useState — matches proto) ──────
-export type BootCtrl = {
-  cancel: () => void;
-  pauseDialog: () => void;
-  resumeDialog: () => void;
-};
-
-// Exported for unit-testing (sysfail-loop.test.ts). Not used by RSC consumers.
-export function runBoot(
-  container: HTMLElement,
-  specs: LinePart[][],
-  dialog: string[],
-  opts: {
-    lineMs: number;
-    lineJitter: number;
-    cmdMs: number;
-    cmdJitter: number;
-    typeMs: number;
-    holdMs: number;
-    backMs: number;
-    interMs: number;
-    startMs: number;
-    onFirstLoop?: () => void;
-  },
-): BootCtrl {
-  let cancelled = false;
-  let dialogPaused = false;
-  let dialogResumeFn: (() => void) | null = null;
-  const timers: ReturnType<typeof setTimeout>[] = [];
-
-  function later(fn: () => void, ms: number) {
-    if (cancelled) return;
-    timers.push(setTimeout(fn, ms));
-  }
-
-  function revealLines(idx: number) {
-    if (cancelled) return;
-    if (idx >= specs.length) {
-      later(typeCmd, 350);
-      return;
-    }
-    const spec = specs[idx];
-    if (!spec) return;
-    container.appendChild(buildLine(spec));
-    later(() => revealLines(idx + 1), opts.lineMs + Math.random() * opts.lineJitter);
-  }
-
-  function typeCmd() {
-    if (cancelled) return;
-    const cmdText = 'run bio.exe --verbose';
-
-    const line = document.createElement('span');
-    line.className = 'boot__line';
-    const prompt = document.createElement('span');
-    prompt.className = 'boot__prompt';
-    prompt.textContent = 'erik@portfolio:~$';
-    const cmdEl = document.createElement('span');
-    cmdEl.className = 'boot__cmd';
-    const cursor = document.createElement('span');
-    cursor.className = 'boot__cursor';
-    line.appendChild(prompt);
-    line.appendChild(document.createTextNode(' '));
-    line.appendChild(cmdEl);
-    line.appendChild(cursor);
-    container.appendChild(line);
-
-    let i = 0;
-    function step() {
-      if (cancelled) return;
-      if (i >= cmdText.length) {
-        cursor.remove();
-        container.appendChild(buildBlankLine());
-        later(startDialog, 350);
-        return;
-      }
-      cmdEl.textContent += cmdText[i++];
-      later(step, opts.cmdMs + Math.random() * opts.cmdJitter);
-    }
-    step();
-  }
-
-  function startDialog() {
-    if (cancelled) return;
-
-    const line = document.createElement('span');
-    line.className = 'boot__line';
-    const prefix = document.createElement('span');
-    prefix.className = 'boot__matrix-prefix';
-    prefix.textContent = '>';
-    const out = document.createElement('span');
-    out.className = 'boot__matrix-out';
-    const cur = document.createElement('span');
-    cur.className = 'boot__cursor';
-    line.appendChild(prefix);
-    line.appendChild(out);
-    line.appendChild(cur);
-    container.appendChild(line);
-
-    let phraseIdx = 0;
-    let charIdx = 0;
-    let phase: 'type' | 'hold' | 'back' = 'type';
-    let firedOnce = false;
-
-    function tick() {
-      if (cancelled) return;
-      // Pause: store self as resume target and yield until resumeDialog() fires.
-      if (dialogPaused) {
-        dialogResumeFn = tick;
-        return;
-      }
-      const phrase = dialog[phraseIdx];
-      if (!phrase) return;
-      if (phase === 'type') {
-        out.textContent = phrase.slice(0, ++charIdx);
-        if (charIdx >= phrase.length) {
-          phase = 'hold';
-          later(tick, opts.holdMs);
-        } else later(tick, opts.typeMs);
-      } else if (phase === 'hold') {
-        phase = 'back';
-        later(tick, opts.backMs);
-      } else {
-        out.textContent = phrase.slice(0, --charIdx);
-        if (charIdx <= 0) {
-          phraseIdx = (phraseIdx + 1) % dialog.length;
-          if (phraseIdx === 0 && !firedOnce) {
-            firedOnce = true;
-            opts.onFirstLoop?.();
-          }
-          phase = 'type';
-          later(tick, opts.interMs);
-        } else {
-          later(tick, opts.backMs);
-        }
-      }
-    }
-    tick();
-  }
-
-  later(() => revealLines(0), opts.startMs);
-
-  return {
-    cancel: () => {
-      cancelled = true;
-      timers.forEach(clearTimeout);
-      dialogResumeFn = null;
-    },
-    pauseDialog: () => {
-      dialogPaused = true;
-    },
-    resumeDialog: () => {
-      dialogPaused = false;
-      if (dialogResumeFn) {
-        const fn = dialogResumeFn;
-        dialogResumeFn = null;
-        fn();
-      }
-    },
-  };
-}
+// Re-export runBoot so sysfail-loop.test.ts can import from this module path
+// without needing to change its import after the extract.
+export { type BootCtrl, runBoot } from '@/lib/boot-animation';
 
 // ── HeroBootAnimation island ──────────────────────────────────────────────────
 // Each variant mounts its own instance; only the one matching the viewport runs.
@@ -245,7 +30,11 @@ type Props = {
 
 export function HeroBootAnimation({ variant }: Props) {
   const bootRef = useRef<HTMLDivElement>(null);
-  const bootCtrl = useRef<BootCtrl | null>(null);
+  // ctrlRef holds the BootCtrl returned by runBoot. Passed into
+  // buildDesktopOnFirstLoop so onFirstLoop accesses it at invocation time
+  // (seconds after runBoot assigns it) instead of capturing a pre-assignment
+  // binding — this eliminates the temporal dead zone (Fix 2).
+  const ctrlRef = useRef<BootCtrl | null>(null);
 
   useEffect(() => {
     if (typeof window === 'undefined') return;
@@ -262,7 +51,8 @@ export function HeroBootAnimation({ variant }: Props) {
     if (!readMotion()) {
       for (const s of specs) el.appendChild(buildLine(s));
       el.appendChild(buildStaticCmdLine());
-      el.appendChild(buildBlankLine());
+      // Fix 5: was buildBlankLine()
+      el.appendChild(buildLine([' ']));
       el.appendChild(buildStaticDialogLine('The Matrix has you...'));
       return;
     }
@@ -291,53 +81,13 @@ export function HeroBootAnimation({ variant }: Props) {
     };
     const timing = variant === 'desktop' ? timingDesktop : timingMobile;
 
-    const onFirstLoop =
-      variant === 'desktop'
-        ? () => {
-            // Find the parent <section> via DOM traversal — works since bootRef
-            // is mounted inside .hero--desktop > .hero__left.
-            const section = el.closest('section');
+    const onFirstLoop = variant === 'desktop' ? buildDesktopOnFirstLoop(el, ctrlRef) : undefined;
 
-            // Skip if hero is off-screen
-            if (section) {
-              const r = section.getBoundingClientRect();
-              if (r.bottom < 0 || r.top > window.innerHeight) return;
-            }
-
-            // 1. Shake hero panel (40ms shake → 40ms shake-2 → 80ms clear)
-            if (section) {
-              section.classList.add('shake');
-              setTimeout(() => section.classList.replace('shake', 'shake-2'), 40);
-              setTimeout(() => section.classList.remove('shake', 'shake-2'), 80);
-            }
-
-            // 2. Pause rain + CRT effects + dialog
-            window.dispatchEvent(new CustomEvent('sysfail:start'));
-            document.documentElement.classList.add('sysfail-on');
-            ctrl.pauseDialog();
-
-            // 3. Signal HeroSystemFailure to show the overlay via DOM event.
-            // HeroSystemFailure listens on 'hero:sysfail:show' and toggles .on.
-            window.dispatchEvent(new CustomEvent('hero:sysfail:show'));
-
-            // 4. After display hold: hide overlay, restore CRT, resume rain + dialog
-            setTimeout(() => {
-              window.dispatchEvent(new CustomEvent('hero:sysfail:hide'));
-              document.documentElement.classList.remove('sysfail-on');
-              // Wait for fade-out transition (300ms) before resuming rain
-              setTimeout(() => {
-                window.dispatchEvent(new CustomEvent('sysfail:end'));
-                ctrl.resumeDialog();
-              }, 300);
-            }, 5000);
-          }
-        : undefined;
-
-    const ctrl = runBoot(el, specs, dialog, onFirstLoop ? { ...timing, onFirstLoop } : timing);
+    ctrlRef.current = runBoot(el, specs, dialog, onFirstLoop ? { ...timing, onFirstLoop } : timing);
 
     const onVisibility = () => {
-      if (document.hidden) ctrl.pauseDialog();
-      else ctrl.resumeDialog();
+      if (document.hidden) ctrlRef.current?.pauseDialog();
+      else ctrlRef.current?.resumeDialog();
     };
     // Visibility listener is only needed for the desktop variant
     // (mobile dialog is short enough that pausing isn't critical)
@@ -345,9 +95,8 @@ export function HeroBootAnimation({ variant }: Props) {
       document.addEventListener('visibilitychange', onVisibility);
     }
 
-    bootCtrl.current = ctrl;
     return () => {
-      ctrl.cancel();
+      ctrlRef.current?.cancel();
       if (variant === 'desktop') {
         document.removeEventListener('visibilitychange', onVisibility);
       }
